requirements.txt
python-telegram-bot
youtube-transcript-api
python-dotenv
openai
